requirements.txt
pycryptodome
pillow
json
base64